🎁 Change set_child_flag transaction to listener
This commit will remove the set_child_flag transaction and replace it
with a listener.

Ref:
  - https://github.com/scientist-softserv/iiif_print/issues/330
